requirements.txt
flask
pyinstaller 
pillow
pystray